Trying to initiate refactoring of difficult location
