+++ washee_app/lib/core/base_usecase/usecase.dart
