+++ washee_app/lib/features/booking/domain/usecases/has_current_booking.dart
@@ -1,5 +1,5 @@
 import 'package:equatable/equatable.dart';
-import 'package:washee/core/standards/base_usecase/usecase.dart';
+import 'package:washee/core/base_usecase/usecase.dart';
 import 'package:washee/features/booking/data/repositories/book_repository_impl.dart';
 
 class HasCurrentBookingUseCase implements UseCase<bool, HasCurrentBookingParams> {
