+++ washee_app/lib/features/location/domain/usecases/connect.dart
@@ -1,5 +1,5 @@
 import 'package:dio/dio.dart';
-import 'package:washee/core/standards/base_usecase/usecase.dart';
+import 'package:washee/core/base_usecase/usecase.dart';
 import 'package:washee/features/location/data/repositories/connect_repo.dart';
 
 class ConnectUseCase implements UseCase<Response, NoParams> {
