+++ washee_app/lib/features/location/domain/usecases/disconnect_box_wifi.dart
@@ -1,4 +1,4 @@
-import 'package:washee/core/standards/base_usecase/usecase.dart';
+import 'package:washee/core/base_usecase/usecase.dart';
 import 'package:washee/features/location/data/repositories/unlock_repo.dart';
 
 class DisconnectBoxWifiUsecase implements UseCase<bool, NoParams> {
