+++ washee_app/lib/features/location/domain/usecases/get_machines.dart
@@ -1,4 +1,4 @@
-import 'package:washee/core/standards/base_usecase/usecase.dart';
+import 'package:washee/core/base_usecase/usecase.dart';
 
 import 'package:washee/features/location/data/models/box_machine_model.dart';
 import 'package:washee/features/location/data/repositories/get_machines_repo.dart';
