+++ washee_app/lib/features/location/domain/usecases/get_wifi_permission.dart
@@ -1,4 +1,4 @@
-import 'package:washee/core/standards/base_usecase/usecase.dart';
+import 'package:washee/core/base_usecase/usecase.dart';
 import 'package:washee/features/location/data/repositories/unlock_repo.dart';
 
 class GetWifiPermissionUsecase implements UseCase<bool, NoParams> {
