+++ washee_app/lib/features/location/presentation/widgets/machine_overview.dart
@@ -5,7 +5,7 @@ import 'package:washee/features/location/data/models/box_machine_model.dart';
 import 'package:washee/injection_container.dart';
 
 import '../../../../core/global_providers/global_provider.dart';
-import '../../../../core/standards/base_usecase/usecase.dart';
+import '../../../../core/base_usecase/usecase.dart';
 import 'machine_card.dart';
 import '../../../../injection_container.dart';
 import '../../domain/usecases/get_machines.dart';
